Java static initializer block
You are given a class Solution with a main method. Complete the given
code so that it outputs the area of a parallelogram with breadth B and
height H . You should read the variables from the standard input.

If B<0or H<0 , the output should be "java.lang.Exception: Breadth and
height must be positive" without quotes.